+++ R/offset_convergence.R
@@ -8,8 +8,3 @@ offset_convergence <- function(r, xind, dgz)
   return(off)
 }
 
-L2 <- function(y, yhat, w){ 
-    n <- length(y)
-    if (missing(w)) w <- rep(1 / n, n)
-    return(stats::weighted.mean((y-yhat)^2, w))
-}