+++ R/smooth_cgam.R
@@ -1,39 +1,3 @@
-
-cgam_lookup <- c(inc = "s.incr", dec = "s.decr", cvx = "s.conv", ccv = "s.conc",
-  inccvx = "s.incr.conv", deccvx = "s.decr.conv", incccv = "s.incr.conc",
-  decccv = "s.decr.conc"
-)
-
-
-cgam.setup <- function(mf, control)
-{
-  # Prepare the formula
-  mt <- attr(mf, "terms")
-  ptot <- length(attr(mt, "term.labels"))
-  gind <- attr(mt, "specials")$g
-  sind <- attr(mt, "specials")$s
-  # Default parameters for control
-  defcontrol <- list()
-  # Match control
-  mc <- names(defcontrol) %in% names(control)
-  control <- c(control, defcontrol[!mc])
-  # Create call to "s" for each term
-  term_s <- sapply(mf[c(gind, sind)], function(x){
-    sfun <- cgam_lookup[attr(x, "fcons")]
-    if (length(sfun) == 0) sfun <- "s"
-    cpars <- list(str2lang(sfun), str2lang(attr(x, "label")))
-    cpars <- c(cpars, attr(x, "s_opts"))
-    cl <- as.call(cpars)
-    deparse(cl)
-  })
-  # Add linear terms
-  term_lin <- attr(mt, "term.labels")[-(c(gind, sind) - 1)]
-  # Put together
-  rhsform <- paste(c(term_s, term_lin), collapse = " + ")
-  control$formula <- stats::reformulate(c(term_s, term_lin), all.vars(mt)[1])
-  return(control)
-}
-
 smooth_cgam <- function(x, y, formula, Xcov, ...){
   cgam_data <- data.frame(as.data.frame(x), Xcov)  
   p <- ncol(cgam_data)
@@ -48,7 +12,7 @@ smooth_cgam <- function(x, y, formula, Xcov, ...){
   colnames(preds) <- c(gfit$xnms, gfit$znms)
   whichinds <- pmatch(colnames(x), colnames(preds))
   gx <- preds[,c(whichinds, setdiff(seq_len(ncol(preds)), whichinds)), drop = FALSE]
-  # Estimate first derivative (different functions for scam or gam)
+  # Estimate first derivative 
   xmat <- cbind(gfit$xmat_add, gfit$zmat)
   colnames(xmat) <- c(gfit$xnms, gfit$znms)
   xmat <- xmat[,c(whichinds, setdiff(seq_len(ncol(preds)), whichinds)), drop = FALSE]